+++ libs/core/src/handlers/events.ts
@@ -24,6 +24,7 @@ export enum SeelenEvent {
   PowerMode = "power-mode",
   BatteriesStatus = "batteries-status",
   ColorsChanged = "colors-changed",
+  SystemFontsChanged = "system::fonts-changed",
   WMSetReservation = "wm::set-reservation",
   WMForceRetiling = "wm::force-retiling",
   WMTreeChanged = "wm::tree-changed",