+++ src/background/modules/mod.rs
@@ -1,5 +1,6 @@
 pub mod apps;
 pub mod clipboard;
+pub mod fonts;
 pub mod media;
 pub mod monitors;
 pub mod network;